Refactor to match adapter-node new file structure
Add start script to example

Format files
