+++ src/index.js
@@ -1,15 +1,15 @@
-import { App } from 'APP';
-import { manifest } from 'MANIFEST';
-import { createServer } from './server.js';
+import { app } from './deps.ts';
+import { handler } from './handler.js';
 
 export const path = Deno.env.get(PATH_ENV) ?? false;
 export const host = Deno.env.get(HOST_ENV) ?? '0.0.0.0';
 export const port = Deno.env.get(PORT_ENV) ?? (!path && 3000);
 
-const app = new App(manifest);
+// TODO: add compression middleware
+const server = app().use(handler);
 
 const addr = path || `${host}:${port}`;
-const instance = createServer(app).listen(addr, (err) => {
+const instance = server.listen(addr, (err) => {
 	if (err) {
 		console.error('error', err);
 	} else {
@@ -17,4 +17,4 @@ const instance = createServer(app).listen(addr, (err) => {
 	}
 });
 
-export { instance };
+export { server, instance };
