+++ src/types.d.ts
@@ -1,0 +1,7 @@
+declare global {
+	const PATH_ENV: string;
+	const HOST_ENV: string;
+	const PORT_ENV: string;
+}
+
+export {};
